+++ CHANGES.rst
@@ -4,6 +4,8 @@
 New Features
 ------------
 
+- Added API for renaming subsets to Subset Tools plugin. [#3356]
+
 Cubeviz
 ^^^^^^^
 
@@ -46,6 +48,8 @@ Bug Fixes
 Cubeviz
 ^^^^^^^
 
+- Fixed copious warnings from spaxel tool when data has INF. [#3368]
+
 Imviz
 ^^^^^
 
@@ -73,6 +77,8 @@ Cubeviz
 Imviz
 ^^^^^
 
+- Spatial subsets no longer show as having mixed visibility (in the legend and plot options tab) when aligned by WCS. [#3373]
+
 Mosviz
 ^^^^^^
 
@@ -92,9 +98,9 @@ New Features
 
 - Improve performance while importing multiple regions. [#3321]
 
-* API method to toggle API hints. [#3336]
+- API method to toggle API hints. [#3336]
 
-* Changing flux/SB display units no longer resets viewer zoom levels. [#3335]
+- Changing flux/SB display units no longer resets viewer zoom levels. [#3335]
 
 Cubeviz
 ^^^^^^^